+++ client/js/component/infiniteScroll/index.less
@@ -1,0 +1,22 @@
+:local(.explore){
+    max-width: 935px;
+    margin: 60px auto 0;
+    .m_row{
+        display: flex;
+        .item{
+            width: 34%;
+            img{
+                width: 100%;
+                height: 293px;
+            }
+            &:nth-child(2){
+                margin:0 30px;
+            }
+        }
+    }
+    .u-infinite{
+        height: 40px;
+        text-align: center;
+        margin: 30px 0;
+    }
+}
