+++ client/js/instagram/explore/index.less
@@ -3,25 +3,7 @@
     margin: 60px auto 0;
     .title{
         color: #999;
-            font-size: 14px;
-            font-weight: 600;
-    }
-    .m_row{
-        display: flex;
-        .item{
-            width: 34%;
-            img{
-                width: 100%;
-                height: 293px;
-            }
-            &:nth-child(2){
-                margin:0 30px;
-            }
-        }
-    }
-    .u-infinite{
-        height: 40px;
-        text-align: center;
-        margin: 30px 0;
+        font-size: 14px;
+        font-weight: 600;
     }
 }
