+++ client/js/instagram/profile/index.less
@@ -44,24 +44,5 @@ body{background-color: #fafafa;}
                 }
         }
     }
-    .m-activity{
-        .ant-tabs-nav-wrap{
-            margin-top: 30px;
-            .ant-tabs-nav{
-                position: relative;
-                left: 50%;
-                transform: translate(-50%,0);
-                .ant-tabs-tab-active{
-                    color:#000;
-                }
-                .ant-tabs-ink-bar{
-                    background-color: #000;
-                }
-                .ant-tabs-tab:hover{
-                    color:#000;
-                }
-            }
-        }
-    }
-    .ant-upload-list-item.ant-upload-list-item-done{display: none;}
+    // .ant-upload-list-item.ant-upload-list-item-done{display: none;}
 }
