+++ client/js/instagram/profile/userPosts/index.less
@@ -1,0 +1,19 @@
+:local(.UserPosts){
+    .ant-tabs-nav-wrap{
+        margin-top: 30px;
+        .ant-tabs-nav{
+            position: relative;
+            left: 50%;
+            transform: translate(-50%,0);
+            .ant-tabs-tab-active{
+                color:#000;
+            }
+            .ant-tabs-ink-bar{
+                background-color: #000;
+            }
+            .ant-tabs-tab:hover{
+                color:#000;
+            }
+        }
+    }
+}
